+++ CHANGELOG.md
@@ -17,7 +17,7 @@ Code Completion:
 1.  Keyword completion support for `ForwardAxis` and `ReverseAxis` for XPath and XQuery.
 1.  Keyword completion support for `KindTest` based `NodeTest`s and `ItemType`s for XPath and XQuery.
 1.  Keyword completion support for `SequenceType`s for XPath and XQuery.
-1.  XML Schema type completion support for `AtomicOrUnionType` and `SimpleTypeName` for XQuery.
+1.  XML Schema type completion support for `AtomicOrUnionType` and `SimpleTypeName` for XPath and XQuery.
 
 Run Configurations: